+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+* Add `--install-toolchain` to install the Rust toolchain specified by `--toolchain` when it is not already installed.
+
 ## [0.7.0] - 2026-08-11
 
 ### Fixed